requirements.txt
python-bitcoinrpc
numpy
click
